rearranging for new save pipeline modal thing
